Move most of torchbench.py code to new utils.py + bert.py example
This commit adds the file `utils.py` to share the logic used by
`torchbench.py` to compile, time, and check the torchbench models with
other models not in torchbench. This commit also adds a script that
runs `Bert` on IREE and PyTorch eager.
